Add workaround for <https://github.com/python/cpython/issues/89553>.
On Python 3.10 and earlier, any `CancelledError` raised inside a task is
rewrapped with another `CancelledError()` without message. (The original
exception is attached as `__context__`.) This swallows the original
cancellation message.

Starting with Python 3.11, `asyncio.Future._make_cancelled_error()` has
another check that ensures the original `CancelledError` is re-used and
no second exception is created.

To bring the two behaviors in line, our custom `Runner` catches the
`CancelledError` and checks if it has another `CancelledError` attached
as `__context__`. If yes, it unwraps the inner exception and reraises
it.